Added ability to underline
Added the ability to underline selected text in EasyMDE.

You add a button with the following options:

name: "underline",
action: EasyMDE.toggleUnderline,
className: "fa fa-underline",
title: "Underline"

This has had to replace the option to use "__" for bold, as that is what various markdown renderers (Showdown, Discord, etc.) use.

It works in exactly the same way as the EasyMDE.toggleBold function.