Added the ability to underline selected text in EasyMDE.
Added the ability to underline selected text in EasyMDE.

You can now add a button with the following options:

name: "underline",
action: EasyMDE.toggleUnderline,
className: "fa fa-underline",
title: "Underline"

This has had to replace the option to use "__" for bold, as that is what various markdown renderers (Showdown etc.) have decided to use.

EasyMDE.toggleUnderline works in exactly the same way as the EasyMDE.toggleBold function, just with "__" instead of "**".